requirements.txt
celery==4.4.7
Django==3.1.8
django-celery-beat==2.1.0
django-cryptography==1.1
django-ratelimit==3.0.0
python-dateutil==2.8.1
redis==3.5.3

gunicorn==20.0.4
whitenoise==5.2.0
psycopg2-binary==2.8.6
dj-database-url==0.5.0
